Rust::com move Lola FFI implementation
* Moved Lola specific FFI implementation in seperate file
