+++ score/mw/com/impl/rust/com-api/com-api-runtime-lola/BUILD
@@ -30,6 +30,7 @@ rust_library(
         "//score/mw/com/impl/rust:mw_com",
         "//score/mw/com/impl/rust/com-api/com-api-concept",
         "//score/mw/com/impl/rust/com-api/com-api-ffi-lola:bridge_ffi_rs",
+        "//score/mw/com/impl/rust/com-api/com-api-ffi-lola:bridge_ffi_lola",
         "@score_communication_crate_index//:futures",
     ],
 )
